services/anomaly-detector/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
scikit-learn
numpy
